requirements.txt
discord.py==2.2.2
python-dotenv==0.16.0
fastapi==0.89.1
uvicorn